+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - `insertAddress` function that calls `EstimateShipping` mutation in order to get available delivery options.
 - `selectDeliveryOption` function that sets the current selected delivery option.
+- Use `useQueueStatus` from `OrderQueue`.
 
 ## [0.1.0] - 2019-09-23